+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -1,15 +1,6 @@
 import { useEffect, useState } from 'react';
-
-import {
-  CallTypes,
-  combineComparators,
-  Comparator,
-  defaultSortPreset,
-  screenSharing,
-  SfuModels,
-  speakerLayoutSortPreset,
-  StreamVideoParticipant,
-} from '@stream-io/video-client';
+import clsx from 'clsx';
+import { SfuModels, StreamVideoParticipant } from '@stream-io/video-client';
 import { useCall, useCallStateHooks } from '@stream-io/video-react-bindings';
 
 import {
@@ -22,16 +13,23 @@ import {
   useHorizontalScrollPosition,
   useVerticalScrollPosition,
 } from '../../../hooks';
-import clsx from 'clsx';
+import { useSpeakerLayoutSortPreset } from './hooks';
+import { useCalculateHardLimit } from '../../hooks/useCalculateHardLimit';
+import { ParticipantsAudio } from '../Audio';
 
 export type SpeakerLayoutProps = {
   ParticipantViewUISpotlight?: ParticipantViewProps['ParticipantViewUI'];
   ParticipantViewUIBar?: ParticipantViewProps['ParticipantViewUI'];
   /**
-   * The position of the particpants who are not in focus.
+   * The position of the participants who are not in focus.
    * Providing `null` will hide the bar.
    */
   participantsBarPosition?: 'top' | 'bottom' | 'left' | 'right' | null;
+  /**
+   * Hard limits the number of the participants rendered in the participants bar.
+   * Providing string `dynamic` will calculate hard limit based on screen width/height.
+   */
+  participantsBarLimit?: 'dynamic' | number;
 } & Pick<ParticipantViewProps, 'VideoPlaceholder'>;
 
 const DefaultParticipantViewUIBar = () => (
@@ -45,45 +43,61 @@ export const SpeakerLayout = ({
   ParticipantViewUISpotlight = DefaultParticipantViewUISpotlight,
   VideoPlaceholder,
   participantsBarPosition = 'bottom',
+  participantsBarLimit,
 }: SpeakerLayoutProps) => {
   const call = useCall();
-  const { useParticipants } = useCallStateHooks();
+  const { useParticipants, useRemoteParticipants } = useCallStateHooks();
   const [participantInSpotlight, ...otherParticipants] = useParticipants();
-  const [scrollWrapper, setScrollWrapper] = useState<HTMLDivElement | null>(
-    null,
+  const remoteParticipants = useRemoteParticipants();
+  const [participantsBarWrapperElement, setParticipantsBarWrapperElement] =
+    useState<HTMLDivElement | null>(null);
+  const [participantsBarElement, setParticipantsBarElement] =
+    useState<HTMLDivElement | null>(null);
+  const [buttonsWrapperElement, setButtonsWrapperElement] =
+    useState<HTMLDivElement | null>(null);
+
+  const isSpeakerScreenSharing = hasScreenShare(participantInSpotlight);
+  const hardLimit = useCalculateHardLimit(
+    buttonsWrapperElement,
+    participantsBarElement,
+    participantsBarLimit,
   );
-  const isOneOnOneCall = otherParticipants.length === 1;
+
+  const isVertical =
+    participantsBarPosition === 'left' || participantsBarPosition === 'right';
+  const isHorizontal =
+    participantsBarPosition === 'top' || participantsBarPosition === 'bottom';
 
   useEffect(() => {
-    if (!scrollWrapper || !call) return;
+    if (!participantsBarWrapperElement || !call) return;
 
-    const cleanup = call.setViewport(scrollWrapper);
+    const cleanup = call.setViewport(participantsBarWrapperElement);
     return () => cleanup();
-  }, [scrollWrapper, call]);
+  }, [participantsBarWrapperElement, call]);
 
-  useEffect(() => {
-    if (!call) return;
-    // always show the remote participant in the spotlight
-    if (isOneOnOneCall) {
-      call.setSortParticipantsBy(combineComparators(screenSharing, loggedIn));
-    } else {
-      call.setSortParticipantsBy(speakerLayoutSortPreset);
-    }
+  const isOneOnOneCall = otherParticipants.length === 1;
+  useSpeakerLayoutSortPreset(call, isOneOnOneCall);
 
-    return () => {
-      // reset the sorting to the default for the call type
-      const callConfig = CallTypes.get(call.type);
-      call.setSortParticipantsBy(
-        callConfig.options.sortParticipantsBy || defaultSortPreset,
-      );
-    };
-  }, [call, isOneOnOneCall]);
+  let participantsWithAppliedLimit = otherParticipants;
+
+  if (typeof participantsBarLimit !== 'undefined') {
+    const hardLimitToApply = isVertical
+      ? hardLimit.vertical
+      : hardLimit.horizontal;
+
+    participantsWithAppliedLimit = otherParticipants.slice(
+      0,
+      // subtract 1 if speaker is sharing screen as
+      // that one is rendered independently from otherParticipants array
+      hardLimitToApply - (isSpeakerScreenSharing ? 1 : 0),
+    );
+  }
 
   if (!call) return null;
 
-  const isSpeakerScreenSharing = hasScreenShare(participantInSpotlight);
   return (
     <div className="str-video__speaker-layout__wrapper">
+      <ParticipantsAudio participants={remoteParticipants} />
       <div
         className={clsx(
           'str-video__speaker-layout',
@@ -95,7 +109,7 @@ export const SpeakerLayout = ({
           {participantInSpotlight && (
             <ParticipantView
               participant={participantInSpotlight}
-              muteAudio={isSpeakerScreenSharing}
+              muteAudio={true}
               trackType={
                 isSpeakerScreenSharing ? 'screenShareTrack' : 'videoTrack'
               }
@@ -104,13 +118,19 @@ export const SpeakerLayout = ({
             />
           )}
         </div>
-        {otherParticipants.length > 0 && participantsBarPosition && (
-          <div className="str-video__speaker-layout__participants-bar-buttons-wrapper">
+        {participantsWithAppliedLimit.length > 0 && participantsBarPosition && (
+          <div
+            ref={setButtonsWrapperElement}
+            className="str-video__speaker-layout__participants-bar-buttons-wrapper"
+          >
             <div
               className="str-video__speaker-layout__participants-bar-wrapper"
-              ref={setScrollWrapper}
+              ref={setParticipantsBarWrapperElement}
             >
-              <div className="str-video__speaker-layout__participants-bar">
+              <div
+                ref={setParticipantsBarElement}
+                className="str-video__speaker-layout__participants-bar"
+              >
                 {isSpeakerScreenSharing && (
                   <div
                     className="str-video__speaker-layout__participant-tile"
@@ -120,10 +140,11 @@ export const SpeakerLayout = ({
                       participant={participantInSpotlight}
                       ParticipantViewUI={ParticipantViewUIBar}
                       VideoPlaceholder={VideoPlaceholder}
+                      muteAudio={true}
                     />
                   </div>
                 )}
-                {otherParticipants.map((participant) => (
+                {participantsWithAppliedLimit.map((participant) => (
                   <div
                     className="str-video__speaker-layout__participant-tile"
                     key={participant.sessionId}
@@ -132,18 +153,21 @@ export const SpeakerLayout = ({
                       participant={participant}
                       ParticipantViewUI={ParticipantViewUIBar}
                       VideoPlaceholder={VideoPlaceholder}
+                      muteAudio={true}
                     />
                   </div>
                 ))}
               </div>
             </div>
-            {(participantsBarPosition === 'left' ||
-              participantsBarPosition === 'right') && (
-              <VerticalScrollButtons scrollWrapper={scrollWrapper} />
+            {isVertical && (
+              <VerticalScrollButtons
+                scrollWrapper={participantsBarWrapperElement}
+              />
             )}
-            {(participantsBarPosition === 'top' ||
-              participantsBarPosition === 'bottom') && (
-              <HorizontalScrollButtons scrollWrapper={scrollWrapper} />
+            {isHorizontal && (
+              <HorizontalScrollButtons
+                scrollWrapper={participantsBarWrapperElement}
+              />
             )}
           </div>
         )}
@@ -222,9 +246,3 @@ const VerticalScrollButtons = <T extends HTMLElement>({
 
 const hasScreenShare = (p?: StreamVideoParticipant) =>
   !!p?.publishedTracks.includes(SfuModels.TrackType.SCREEN_SHARE);
-
-const loggedIn: Comparator<StreamVideoParticipant> = (a, b) => {
-  if (a.isLocalParticipant) return 1;
-  if (b.isLocalParticipant) return -1;
-  return 0;
-};