+++ sample-apps/react/react-video-demo/src/utils/useGenerateUser.ts
@@ -1,4 +1,4 @@
-import { uniqueNamesGenerator, Config, starWars } from 'unique-names-generator';
+import { Config, starWars, uniqueNamesGenerator } from 'unique-names-generator';
 import { v1 as uuid } from 'uuid';
 import { getURLCredentials } from './getURLCredentials';
 
@@ -18,13 +18,13 @@ const config: Config = {
 
 export const generateUser = (): User => {
   const { user_id, user_name, id } = getURLCredentials();
-  const characterName: string = uniqueNamesGenerator(config);
+  const characterName = uniqueNamesGenerator(config);
 
-  const userId: string = user_id ?? `demo-${uuid()}`;
-  const userName: string = user_name ?? `${characterName}`;
+  const userName = user_name ?? characterName;
+  const userId = user_id ?? `demo-${userName}-${uuid()}`;
 
   return {
-    id: userId,
+    id: userId.replace(/[^_\-0-9a-zA-Z@]/g, '_'),
     name: userName,
     role: id ? 'user' : 'admin',
     teams: ['@stream-io/video-demo'],