+++ packages/i18n/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.1.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.1.1...@stream-io/i18n-0.1.2) (2023-09-25)
+
+
+### Bug Fixes
+
+* Add extra delay before attempting to play video in Safari and Firefox ([#1106](https://github.com/GetStream/stream-video-js/issues/1106)) ([5b4a589](https://github.com/GetStream/stream-video-js/commit/5b4a58918240a7b63807726609d6d54b92cfe1d2))
+
 ### [0.1.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.1.0...@stream-io/i18n-0.1.1) (2023-07-17)