+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.22](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.21...@stream-io/video-react-native-sdk-0.0.22) (2023-09-26)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.25`
+### [0.0.21](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.20...@stream-io/video-react-native-sdk-0.0.21) (2023-09-26)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.24`
+### [0.0.20](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.19...@stream-io/video-react-native-sdk-0.0.20) (2023-09-25)
+
+### Dependency Updates
+
+* `@stream-io/i18n` updated to version `0.1.2`
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.23`
+
+### Bug Fixes
+
+* Add extra delay before attempting to play video in Safari and Firefox ([#1106](https://github.com/GetStream/stream-video-js/issues/1106)) ([5b4a589](https://github.com/GetStream/stream-video-js/commit/5b4a58918240a7b63807726609d6d54b92cfe1d2))
+
 ### [0.0.19](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.18...@stream-io/video-react-native-sdk-0.0.19) (2023-09-20)
 
 ### Dependency Updates